performance-tester/requirements.txt
# Web Framework
Flask==3.0.0
Flask-SocketIO==5.3.5
Flask-CORS==4.0.0

# HTTP Clients & Load Testing
requests==2.31.0
aiohttp==3.9.1
httpx==0.25.2
locust==2.20.0

# System Monitoring
psutil==5.9.6
py-cpuinfo==9.0.0

# Data Processing & Visualization
pandas==2.1.4
numpy==1.26.2
plotly==5.18.0
matplotlib==3.8.2
seaborn==0.13.0

# Database
SQLAlchemy==2.0.23

# Utils
python-dotenv==1.0.0
pydantic==2.5.3
pydantic-settings==2.1.0
python-dateutil==2.8.2
pytz==2023.3

# WebSocket
python-socketio==5.10.0
eventlet==0.33.3

# Report Generation
jinja2==3.1.2
weasyprint==60.1

# CLI & Progress
tqdm==4.66.1
colorama==0.4.6
rich==13.7.0

# Statistics
scipy==1.11.4
